Refactor datafile Python package: add subpackages
- Create datafile.scraper subpackage
- Create datafile.scraper.parsers subpackage
- Modify pipeline analysis and tests accordingly
