+++ src/python/ensembl/production/datafile/scraper/parsers/file_parser.py
@@ -9,15 +9,15 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
+"""Base File Parser for DataFile Scraper package"""
 
 from dataclasses import dataclass
 from datetime import datetime
 from pathlib import Path
-import re
-from typing import Dict, List, Tuple, Optional, Type, Match
+from typing import Dict, List, Tuple, Optional
 from urllib.parse import urljoin
 
-from .utils import blake2bsum
+from ..utils import blake2bsum, make_release
 
 
 FILE_COMPRESSIONS = {"gz": "gzip"}
@@ -81,44 +81,9 @@ class FileMetadata:
 
 
 @dataclass
-class EMBLOptMetadata(BaseOptMetadata):
-    compression: Optional[str]
-    content_type: Optional[str]
-    sorting: Optional[str]
-
-
-@dataclass
-class BAMOptMetadata(BaseOptMetadata):
-    source: Optional[str]
-    origin: Optional[str]
-
-
-@dataclass
-class FASTAOptMetadata(BaseOptMetadata):
-    compression: Optional[str]
-    content_type: Optional[str]
-    sequence_type: str
-
-
-@dataclass
 class Result:
     file_metadata: Optional[FileMetadata]
     errors: List[str]
-
-
-def get_group(
-    group_name: str, match: Optional[Match], default: Optional[str] = None
-) -> Optional[str]:
-    if match:
-        try:
-            return match.group(group_name)
-        except IndexError:
-            pass
-    return default
-
-
-def make_release(ens_version: int) -> Tuple[int, int]:
-    return ens_version, ens_version - 53
 
 
 class FileParserError(ValueError):
@@ -219,118 +184,3 @@ class FileParser(BaseFileParser):
             species=metadata["species"].lower(),
             ens_release=int(metadata["ens_release"]),
         )
-
-
-class EMBLFileParser(FileParser):
-    FILENAMES_RE = re.compile(
-        (
-            r"^(?P<species>\w+)\.(?P<assembly>[\w\-\.]+?)\.(5|1)\d{1,2}\."
-            r"(?P<content_type>abinitio|chr|chr_patch_hapl_scaff|nonchromosomal|(chromosome|plasmid|scaffold)\.[\w\-\.]+?|primary_assembly[\w\-\.]*?|(chromosome_)?group\.\w+?)?.*?"
-        )
-    )
-    FILE_EXT_RE = re.compile(
-        (
-            r".*?\.(?P<file_extension>gff3|gtf|dat)"
-            r"(\.(?P<sorted>sorted))?(\.(?P<compression>gz))?$"
-        )
-    )
-
-    def get_optional_metadata(self, metadata: dict) -> EMBLOptMetadata:
-        match = self.FILENAMES_RE.match(metadata["file_name"])
-        matched_content_type = get_group("content_type", match)
-        match = self.FILE_EXT_RE.match(metadata["file_name"])
-        matched_compression = FILE_COMPRESSIONS.get(get_group("compression", match))
-        file_extension = get_group("file_extension", match)
-        matched_sorting = get_group("sorted", match)
-        compression = (
-            metadata.get("extras", {}).get("compression") or matched_compression
-        )
-        content_type = (
-            metadata.get("extras", {}).get("content_type") or matched_content_type
-        )
-        sorting = metadata.get("extras", {}).get("sorting") or matched_sorting
-        optional_data = EMBLOptMetadata(
-            compression=compression,
-            file_extension=file_extension,
-            content_type=content_type,
-            sorting=sorting,
-        )
-        return optional_data
-
-
-class FASTAFileParser(FileParser):
-    FILENAMES_RE = re.compile(
-        (
-            r"^(?P<species>\w+)\.(?P<assembly>[\w\-\.]+?)\.(?P<sequence_type>dna(_sm|_rm)?|cdna|cds|pep|ncrna)\."
-            r"(?P<content_type>abinitio|all|alt|toplevel|nonchromosomal|(chromosome|plasmid|scaffold)\.[\w\-\.]+?|primary_assembly[\w\-\.]*?|(chromosome_)?group\.\w+?)?.*?"
-        )
-    )
-    FILE_EXT_RE = re.compile(
-        r".*?\.(?P<file_extension>fa)(\.(?P<compression>gz)?(\.gzi|\.fai)?)?$"
-    )
-
-    def get_optional_metadata(self, metadata: dict) -> FASTAOptMetadata:
-        match = self.FILENAMES_RE.match(metadata["file_name"])
-        matched_sequence_type = get_group("sequence_type", match)
-        matched_content_type = get_group("content_type", match)
-        match = self.FILE_EXT_RE.match(metadata["file_name"])
-        file_extension = get_group("file_extension", match)
-        matched_compression = FILE_COMPRESSIONS.get(get_group("compression", match))
-        compression = (
-            metadata.get("extras", {}).get("compression") or matched_compression
-        )
-        content_type = (
-            metadata.get("extras", {}).get("content_type") or matched_content_type
-        )
-        sequence_type = (
-            metadata.get("extras", {}).get("sequence_type") or matched_sequence_type
-        )
-        optional_data = FASTAOptMetadata(
-            compression=compression,
-            file_extension=file_extension,
-            content_type=content_type,
-            sequence_type=sequence_type,
-        )
-        return optional_data
-
-
-class BAMFileParser(FileParser):
-    FILENAMES_RE = re.compile(
-        r"^(?P<assembly>.+?)(\.(?P<source>[a-zA-Z]{2,}))?\.(?P<origin>[\w\-]+)(\.\d)?.*?"
-    )
-    FILE_EXT_RE = re.compile(r".*?\.(?P<file_extension>(bam(\..{2,})?)|txt)$")
-
-    def get_optional_metadata(self, metadata: dict) -> BAMOptMetadata:
-        match = self.FILENAMES_RE.match(metadata["file_name"])
-        matched_source = get_group("source", match)
-        matched_origin = get_group("origin", match)
-        match = self.FILE_EXT_RE.match(metadata["file_name"])
-        file_extension = get_group("file_extension", match)
-        source = metadata.get("extras", {}).get("source") or matched_source
-        origin = metadata.get("extras", {}).get("origin") or matched_origin
-        optional_data = BAMOptMetadata(
-            file_extension=file_extension,
-            source=source,
-            origin=origin,
-        )
-        return optional_data
-
-
-PARSERS = {
-    "embl": EMBLFileParser,
-    "genbank": EMBLFileParser,
-    "gff3": EMBLFileParser,
-    "gtf": EMBLFileParser,
-    "fasta": FASTAFileParser,
-    "bamcov": BAMFileParser,
-}
-
-
-def get_parser(
-    file_format: str, file_path: str
-) -> Tuple[Optional[Type[FileParser]], Optional[Result]]:
-    ParserClass = PARSERS.get(file_format)
-    if ParserClass is None:
-        err = f"Invalid file_format: {file_format} for {file_path}"
-        return None, Result(None, [err])
-    return ParserClass, None
