+++ src/python/ensembl/production/datafile/scraper/serializers.py
@@ -1,0 +1,34 @@
+#  See the NOTICE file distributed with this work for additional information
+#  regarding copyright ownership.
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#      http://www.apache.org/licenses/LICENSE-2.0
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+"""Serializers for DataFile Scraper"""
+
+from pathlib import Path
+from typing import Optional, Tuple
+
+from .utils import get_metadata_from_db, ManifestRow
+
+
+def metadata_from_manifest(manifest_row: ManifestRow, manifest_dir: Path):
+    data = manifest_row._asdict()
+    data["file_dir"] = str(manifest_dir)
+    return data
+
+
+def metadata_from_db(
+    metadata_db_url: str, species: str, ens_release: int
+) -> Tuple[Optional[dict], Optional[str]]:
+    ens_metadatas, err = get_metadata_from_db(metadata_db_url, species, ens_release)
+    if err:
+        return None, err
+    data = ens_metadatas[0]._asdict()
+    data["release_date"] = data["release_date"].isoformat()
+    return data, None
