+++ src/python/ensembl/production/hive/BaseProdRunnable.py
@@ -9,6 +9,7 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
+"""Base Production Python eHive Runnable"""
 
 import json
